Update First Player Tab to Tools Tab with Multiple Options
